feat: extend delegate idea to simple instant interactions
