--- src/main/kotlin/io/github/hytalekt/kytale/interaction/KyInteractionExecutorHolder.kt
@@ -1,4 +1,0 @@
-package io.github.hytalekt.kytale.interaction
-
-class KyInteractionExecutorHolder<Lambda>(var executor: Lambda? = null)
-
