+++ src/main/kotlin/io/github/hytalekt/kytale/interaction/simpleblock/KytaleSimpleBlockInteractionDsl.kt
@@ -1,3 +1,5 @@
+@file:Suppress("NOTHING_TO_INLINE")
+
 package io.github.hytalekt.kytale.interaction.simpleblock
 
 import com.hypixel.hytale.server.core.modules.interaction.interaction.config.client.SimpleBlockInteraction
@@ -28,3 +30,9 @@ inline fun internalSimpleBlockInteraction(
         codec = codecBuilder.build(),
     )
 }
+
+// Create a new anonymous class + anonymous object to delegate the firstRun call.
+// Since this function is inline, a new class will be created for each call, circumventing the
+// restriction of one class per interactionId.
+inline fun createDelegatedSimpleBlockInteraction(): KytaleSimpleBlockInteractionDelegate =
+    object : KytaleSimpleBlockInteractionDelegate() {}
