+++ src/main/kotlin/io/github/hytalekt/kytale/interaction/simpleinstant/KytaleSimpleInstantInteractionDsl.kt
@@ -4,33 +4,31 @@ package io.github.hytalekt.kytale.interaction.simpleinstant
 
 import com.hypixel.hytale.server.core.modules.interaction.interaction.config.SimpleInstantInteraction
 import io.github.hytalekt.kytale.codec.CodecBuilder
-import io.github.hytalekt.kytale.interaction.KyInteractionExecutorHolder
 import io.github.hytalekt.kytale.interaction.KytaleInteractionBundle
 import io.github.hytalekt.kytale.interaction.KytaleInteractionDsl
 import io.github.hytalekt.kytale.codec.newCodecBuilder
 
 inline fun internalSimpleInstantInteraction(
-    interactionId: String,
-    block: @KytaleInteractionDsl KytaleSimpleInstantInteractionBuilder.() -> Unit
+    interactionId: String, block: @KytaleInteractionDsl KytaleSimpleInstantInteractionBuilder.() -> Unit
 ): KytaleInteractionBundle<SimpleInstantInteraction> {
-    val firstRunHolder = KyInteractionExecutorHolder<FirstRunExecutor>()
-    val simpleInstantInteraction = createDelegatedSimpleInstantInteraction(
-        interactionId = interactionId,
-        firstRunHolder = firstRunHolder,
-    )
+    val delegatedInteraction = createDelegatedSimpleInstantInteraction()
 
     val codecBuilder = newCodecBuilder(
-        parentCodec = SimpleInstantInteraction.CODEC,
-        supplier = { simpleInstantInteraction })
+        parentCodec = SimpleInstantInteraction.CODEC, supplier = { delegatedInteraction as SimpleInstantInteraction })
 
     KytaleSimpleInstantInteractionBuilder(
-        codecBuilderScope = CodecBuilder(codecBuilder),
-        firstRunHolder = firstRunHolder,
+        codecBuilderScope = CodecBuilder(codecBuilder), delegate = delegatedInteraction
     ).apply(block)
 
     return KytaleInteractionBundle(
         id = interactionId,
-        interactionClass = simpleInstantInteraction.javaClass,
+        interactionClass = delegatedInteraction.javaClass,
         codec = codecBuilder.build(),
     )
 }
+
+// Create a new anonymous class + anonymous object to delegate all calls.
+// Since this function is inline, a new class will be created for each call,
+// circumventing the restriction of one class per interactionId.
+inline fun createDelegatedSimpleInstantInteraction(): KytaleSimpleInstantInteractionDelegate =
+    object : KytaleSimpleInstantInteractionDelegate() {}
